+++ DEV/trunk/org.openl.core/src/org/openl/types/impl/IBenchmarkableMethod.java
@@ -25,7 +25,7 @@ public interface IBenchmarkableMethod extends IOpenMethod {
 
     String getBenchmarkName();
 
-    Object invokeBenchmark(Object target, Object[] params, IRuntimeEnv env, int ntimes);
+    Object invokeBenchmark(Object target, Object[] params, IRuntimeEnv env, long ntimes);
 
     int nUnitRuns();
